feat: add reservation creation flow
